backend/requirements.txt
# FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Document Processing
Pillow==10.1.0
pytesseract==0.3.10
python-multipart==0.0.6
PyPDF2==3.0.1
# pdf2image==1.16.3  # Optional - not needed as Gemini handles PDFs natively

# AI/LLM Integration - Gemini
google-generativeai==0.3.2

# Database - Supabase
supabase==2.10.0
psycopg2-binary==2.9.9

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.26.0
